+++ grid2op/Agent/__init__.py
@@ -1,5 +1,5 @@
 __all__ = [
-    "Agent", 
+    "BaseAgent",
     "DoNothingAgent",
     "OneChangeThenNothing",
     "GreedyAgent",
@@ -7,10 +7,11 @@ __all__ = [
     "TopologyGreedy",
     "AgentWithConverter",
     "RandomAgent",
-    "MLAgent"
+    "MLAgent",
+    "Agent"
 ]
 
-from grid2op.Agent.Agent import Agent
+from grid2op.Agent.BaseAgent import BaseAgent
 from grid2op.Agent.DoNothing import DoNothingAgent
 from grid2op.Agent.OneChangeThenNothing import OneChangeThenNothing
 from grid2op.Agent.GreedyAgent import GreedyAgent
@@ -19,4 +20,12 @@ from grid2op.Agent.TopologyGreedy import TopologyGreedy
 from grid2op.Agent.AgentWithConverter import AgentWithConverter
 from grid2op.Agent.RandomAgent import RandomAgent
 from grid2op.Agent.MLAgent import MLAgent
+import warnings
 
+
+class TODOLATERAgent(BaseAgent):
+    def __init__(self, *args, **kwargs):
+        BaseAgent.__init__(*args, **kwargs)
+        warnings.warn("Agent class has been renamed \"BaseAgent\". The Agent class will be removed"
+                      "in future versions.",
+                      category=PendingDeprecationWarning)