+++ grid2op/Converter/Converters.py
@@ -13,7 +13,7 @@ This can be done easily with the :class:`IdToAct` class.
 More concretely, the diagram of an agent is:
 
 i) receive an observation (in a form of an object of class :class:`grid2op.BaseObservation.BaseObservation`)
-ii) implement the :func:`grid2op.Agent.Agent.act` taking as input an :class:`grid2op.BaseObservation.BaseObservation` and
+ii) implement the :func:`grid2op.BaseAgent.BaseAgent.act` taking as input an :class:`grid2op.BaseObservation.BaseObservation` and
     returning an :class:`grid2op.BaseAction.BaseAction`
 iii) this :class:`grid2op.BaseAction.BaseAction` is then digested by the environment
 
@@ -21,7 +21,7 @@ Introducing some converters lead to the following:
 
 i) receive an observation (:class:`grid2op.BaseObservation.BaseObservation`)
 ii) the transformer automatically (using :func:`Converter.convert_obs`) to a `transformed observation`
-iii) implement the function :func:`grid2op.Agent.AgentWithConverter.my_act` that takes as input
+iii) implement the function :func:`grid2op.BaseAgent.AgentWithConverter.my_act` that takes as input
      a `transformed observation` and returns an `encoded action`
 iv) the transformer automatically transforms back the `encoded action` into a proper :class:`grid2op.BaseAction.BaseAction`
 v) this :class:`grid2op.BaseAction.BaseAction` is then digested by the environment
@@ -30,7 +30,7 @@ This simple mechanism allows people to focus on iii) above (typically implemente
 without having to worry each time about the complex representations of actions and observations.
 
 More details and a concrete example is given in the documentation of the class
-:class:`grid2op.Agent.AgentWithConverter`.
+:class:`grid2op.BaseAgent.AgentWithConverter`.
 
 Some examples of converters are given in :class:`IdToAct` and :class:`ToVect`.
 """