+++ grid2op/EpisodeData.py
@@ -14,7 +14,7 @@ with:
     - "grid_path": the path where the powergrid has been loaded from
 
   - "episode_times.json": gives some information about the total time spend in multiple part of the runner, mainly the
-    :class:`grid2op.Agent` (and especially its method :func:`grid2op.Agent.act`) and amount of time spent in the
+    :class:`grid2op.BaseAgent` (and especially its method :func:`grid2op.BaseAgent.act`) and amount of time spent in the
     :class:`grid2op.Environment`
 
   - "_parameters.json": is a representation as json of a the :class:`grid2op.Parameters.Parameters` used for this episode
@@ -22,19 +22,19 @@ with:
     reward at index `i` is the one observed by the agent at time `i` and **NOT** the reward sent by the
     :class:`grid2op.Environment` after the action has been implemented.
   - "exec_times.npy" is a numpy 1d array giving the execution time of each time step of the episode
-  - "actions.npy" gives the actions that has been taken by the :class:`grid2op.Agent.Agent`. At row `i` of "actions.npy" is a
+  - "actions.npy" gives the actions that has been taken by the :class:`grid2op.BaseAgent.BaseAgent`. At row `i` of "actions.npy" is a
     vectorized representation of the action performed by the agent at timestep `i` *ie.* **after** having observed
     the observation present at row `i` of "observation.npy" and the reward showed in row `i` of "rewards.npy".
   - "disc_lines.npy" gives which lines have been disconnected during the simulation of the cascading failure at each
     time step. The same convention as for "rewards.npy" has been adopted. This means that the powerlines are
-    disconnected when the :class:`grid2op.Agent` takes the :class:`grid2op.BaseAction` at time step `i`.
+    disconnected when the :class:`grid2op.BaseAgent` takes the :class:`grid2op.BaseAction` at time step `i`.
   - "observations.npy" is a numpy 2d array reprensenting the :class:`grid2op.BaseObservation.BaseObservation` at the disposal of the
-    :class:`grid2op.Agent` when he took his action.
+    :class:`grid2op.BaseAgent` when he took his action.
   - "env_modifications.npy" is a 2d numpy array representing the modification of the powergrid from the environment.
     these modification usually concerns the hazards, maintenance, as well as modification of the generators production
     setpoint or the loads consumption.
 
-All of the above should allow to read back, and better understand the behaviour of some :class:`grid2op.Agent.Agent`, even
+All of the above should allow to read back, and better understand the behaviour of some :class:`grid2op.BaseAgent.BaseAgent`, even
 though such utility functions have not been coded yet.
 """
 
@@ -263,8 +263,8 @@ class EpisodeData:
                 env._time_powerflow)
             self.episode_times["Env"]["observation_computation"] = float(
                 env._time_extract_obs)
-            self.episode_times["Agent"] = {}
-            self.episode_times["Agent"]["total"] = float(time_act)
+            self.episode_times["BaseAgent"] = {}
+            self.episode_times["BaseAgent"]["total"] = float(time_act)
             self.episode_times["total"] = float(end_ - beg_)
 
     def to_disk(self):