+++ grid2op/Observation/BaseObservation.py
@@ -1,7 +1,7 @@
 """
-In a "reinforcement learning" framework, an :class:`grid2op.Agent` receive two information before taking any action on
+In a "reinforcement learning" framework, an :class:`grid2op.BaseAgent` receive two information before taking any action on
 the :class:`grid2op.Environment`. One of them is the :class:`grid2op.Reward` that tells it how well the past action
-performed. The second main input received from the environment is the :class:`BaseObservation`. This is gives the Agent
+performed. The second main input received from the environment is the :class:`BaseObservation`. This is gives the BaseAgent
 partial, noisy, or complete information about the current state of the environment. This module implement a generic
 :class:`BaseObservation`  class and an example of a complete observation in the case of the Learning
 To Run a Power Network (`l2RPN <https://l2rpn.chalearn.org/>`_ ) competition.
@@ -18,11 +18,11 @@ This forecasted powergrid used:
 This functionality was originally attached to the Environment and could only be used to simulate the effect of an action
 on this unique time step. We wanted in this recoding to change that:
 
-  - in an RL setting, an :class:`grid2op.Agent` should not be able to look directly at the :class:`grid2op.Environment`.
-    The only information about the Environment the Agent should have is through the :class:`grid2op.BaseObservation` and
+  - in an RL setting, an :class:`grid2op.BaseAgent` should not be able to look directly at the :class:`grid2op.Environment`.
+    The only information about the Environment the BaseAgent should have is through the :class:`grid2op.BaseObservation` and
     the :class:`grid2op.Reward`. Having this principle implemented will help enforcing this principle.
   - In some wider context, it is relevant to have these forecasts available in multiple way, or modified by the
-    :class:`grid2op.Agent` itself (for example having forecast available for the next 2 or 3 hours, with the Agent able
+    :class:`grid2op.BaseAgent` itself (for example having forecast available for the next 2 or 3 hours, with the BaseAgent able
     not only to change the topology of the powergrid with actions, but also the injections if he's able to provide
     more accurate predictions for example.
 
@@ -180,7 +180,7 @@ class BaseObservation(GridObjects):
 
             - there is a `0`: the powerline is not disconnected from the grid for maintenance
             - there is a `1`, `2`, ... the powerline will be disconnected for at least `1`, `2`, ... timestep (**NB**
-              in all case, the powerline will stay disconnected until a :class:`grid2op.Agent.Agent` performs the
+              in all case, the powerline will stay disconnected until a :class:`grid2op.BaseAgent.BaseAgent` performs the
               proper :class:`grid2op.BaseAction.BaseAction` to reconnect it).
 
     target_dispatch: :class:`numpy.ndarray`, dtype:float